Refactored launch_server into a utility
